requirements.txt
python_telegram_bot==5.2.0
